requirements.txt
acryl_datahub==0.8.41.1
clickhouse_driver==0.2.3
DingtalkChatbot==1.5.3
loguru==0.6.0
